# experimental/gateway/openapi.json
{
  "components": {
    "schemas": {
...
      "ArgCheck": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why dimos wouldn't take it (its own message), else null",
            "examples": [
              "Unknown blueprint configuration option --bogus."
            ],
            "title": "Error"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What it sets: `g.<key>` (GlobalConfig), `<module>.<field>`, `transports.<...>`, or `run --<option>` (one of `dimos run`'s own); null when dimos doesn't know it",
            "examples": [
              "g.robot-ip"
            ],
            "title": "Target"
          },
          "tokens": {
            "description": "The argv items this option took (itself and its value, when given apart)",
            "examples": [
              [
                "--robot-ip",
                "10.0.0.2"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Tokens",
            "type": "array"
          }
        },
        "required": [
          "tokens",
          "target",
          "error"
        ],
        "title": "ArgCheck",
        "type": "object"
      },
      "ArgsCheck": {
        "additionalProperties": true,
        "properties": {
          "args": {
            "description": "Each option, in order",
            "items": {
              "$ref": "#/components/schemas/ArgCheck"
            },
            "title": "Args",
            "type": "array"
          },
          "name": {
            "description": "The blueprint",
            "examples": [
              "unitree-go2"
            ],
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "args"
        ],
        "title": "ArgsCheck",
        "type": "object"
      },
      "ArgsCheckRequest": {
        "additionalProperties": true,
        "properties": {
          "args": {
            "description": "`dimos run <name>` arguments, one argv item each",
            "examples": [
              [
                "--simulation=mujoco"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Args",
            "type": "array"
          }
        },
        "required": [
          "args"
        ],
        "title": "ArgsCheckRequest",
        "type": "object"
      },
...
      "BlueprintConfig": {
        "additionalProperties": true,
        "properties": {
          "modules": {
            "description": "Each module's configurable args",
            "items": {
              "$ref": "#/components/schemas/ModuleConfig"
            },
            "title": "Modules",
            "type": "array"
          },
          "name": {
            "description": "The blueprint",
            "examples": [
              "unitree-go2"
            ],
            "title": "Name",
            "type": "string"
          },
          "overrides": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "description": "Desktop's saved module config for this blueprint (config.yaml `dimos.module_config.<name>`), `--<module>.<field>=value` on every launch of it; secrets as \u2022\u2022\u2022",
            "examples": [
              {
                "go2connection": {
                  "lidar": false
                }
              }
            ],
            "title": "Overrides",
            "type": "object"
          }
        },
        "required": [
          "name",
          "modules",
          "overrides"
        ],
        "title": "BlueprintConfig",
        "type": "object"
      },
      "BlueprintConfigUpdate": {
        "additionalProperties": true,
        "properties": {
          "overrides": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "description": "The blueprint's new saved module config, {module: {field: value}}, replacing the old one; null drops a field; \u2022\u2022\u2022 keeps a saved secret",
            "examples": [
              {
                "go2connection": {
                  "lidar": false
                },
                "voxelgridmapper": {
                  "voxel_size": 0.1
                }
              }
            ],
            "title": "Overrides",
            "type": "object"
          }
        },
        "required": [
          "overrides"
        ],
        "title": "BlueprintConfigUpdate",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": true,
        "description": "Every error answer (4xx and 5xx).",
        "properties": {
          "error": {
            "description": "What went wrong, in words a person can act on",
            "examples": [
              "no upload u7"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "LaunchRequest": {
        "additionalProperties": true,
        "properties": {
          "args": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "`dimos run <blueprint>` arguments, exactly as on its command line (one argv item each, never a shell string): GlobalConfig flags (`--simulation=mujoco`, `--robot-ip 10.0.0.2`, `--no-rerun-web`), module fields (`--voxel-size=0.1`, `--go2connection.lidar=false`), `--disable <module>`, `--config <file>`. They go last, so they win over Desktop's saved config, `overrides` and `replay`. Checked first as POST /dimos/blueprints/{name}/args does: 400 naming each one dimos wouldn't take (`--daemon` and `--help` included). Kept for POST /dimos/runs/restart. New in API 1.17.",
            "examples": [
              [
                "--simulation=mujoco"
              ]
            ],
            "title": "Args"
          },
          "blueprint": {
            "description": "blueprint name",
            "examples": [
              "unitree-go2"
            ],
            "title": "Blueprint",
            "type": "string"
          },
          "overrides": {
            "description": "This launch's own config, on top of Desktop's saved config and never saved: `{global?, modules?, secrets?}`, or (the older form) a flat `{GlobalConfig key: value}`. null drops a saved value for this launch; \u2022\u2022\u2022 keeps a saved secret. Checked against dimos's schemas: 400 with the path and why. Secret values (by name, or listed) go to `dimos run` as environment variables, never argv.",
            "examples": [
              {
                "global": {
                  "n_workers": 4
                },
                "modules": {
                  "go2connection": {
                    "lidar": false
                  }
                }
              },
              {
                "simulation": "mujoco"
              }
            ],
            "properties": {
              "global": {
                "description": "{GlobalConfig key: value}",
                "type": "object"
              },
              "modules": {
                "description": "{module (its name in the blueprint): {field: value}}",
                "type": "object"
              },
              "secrets": {
                "description": "More paths to treat as secrets (`robot_ip`, `<module>.<field>`)",
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            },
            "type": "object"
          },
          "replay": {
            "default": false,
            "description": "Run on a recording: adds `--replay`",
            "title": "Replay",
            "type": "boolean"
          }
        },
        "required": [
          "blueprint"
        ],
        "title": "LaunchRequest",
        "type": "object"
      },
...
      "ModuleConfig": {
        "additionalProperties": true,
        "properties": {
          "args": {
            "description": "Its configurable args (empty when `error`)",
            "items": {
              "$ref": "#/components/schemas/ConfigArg"
            },
            "title": "Args",
            "type": "array"
          },
          "class": {
            "description": "Its Python class",
            "title": "Class",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only when its config couldn't be read: why",
            "examples": [
              "TypeError: CameraModule.config (dict) isn't a pydantic model"
            ],
            "title": "Error"
          },
          "module": {
            "description": "The module's name in the blueprint",
            "examples": [
              "camera"
            ],
            "title": "Module",
            "type": "string"
          }
        },
        "required": [
          "module",
          "class",
          "args"
        ],
        "title": "ModuleConfig",
        "type": "object"
      },
...
      "RunList": {
        "additionalProperties": true,
        "properties": {
          "launch": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Launch"
              },
              {
                "type": "null"
              }
            ],
            "description": "The launch this gateway started (null: none yet)"
          },
          "runs": {
            "description": "Live runs, newest first",
            "items": {
              "$ref": "#/components/schemas/RegistryRun"
            },
            "title": "Runs",
            "type": "array"
          },
          "seenOnBus": {
            "description": "dimos runs heard on the bus that no run in `runs` accounts for (on another machine, or not visible here); they can't be stopped from here",
            "items": {
              "$ref": "#/components/schemas/BusRun"
            },
            "title": "Seenonbus",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "launch"
        ],
        "title": "RunList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The dimos gateway's HTTP API: blueprints, global config, runs and their logs, events, Dimensional cloud uploads,\ndiscovery (blueprints, modules, message types), docs, extras and jobs.\n\ndimOS Desktop starts the gateway (`python -m experimental.gateway`, on a unix socket) and forwards `/dimos/...` to it unchanged.\nEvery answer is JSON unless the operation says otherwise; every error is `{\"error\": \"<message>\"}` (ErrorResponse).\nState is HTTP; changes are events, published on zenoh at `<ns>/dimos/events/<type>` (the `events` tag).\n\nExtensions: `x-family: dimos` on every operation; `x-agent: true` marks what Desktop's agent finds through\nsearch_endpoints; `x-mcp-tool` names an MCP tool that does the same; `x-zenoh-key` on an event schema is its key.\n",
    "title": "dimos gateway",
    "version": "1.18.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dimos/blueprints/{name}/args": {
      "post": {
        "description": "Reads `args` (argv items, never a shell string) with dimos's own blueprint config parser, in a child process that imports the blueprint (cached 10 min per blueprint and args): each option on its own, with the items it took, what it sets (`g.<key>`, `<module>.<field>`, `run --<option>`) and dimos's error for it. `--daemon` and `--help` are refused (a launch runs in the foreground), and so is anything but options (one blueprint per launch). POST /dimos/runs checks its `args` the same way. No side effects. New in API 1.17.",
        "operationId": "post_dimos_blueprints_name_args",
        "parameters": [
          {
            "description": "blueprint name, e.g. unitree-go2-basic",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "description": "blueprint name, e.g. unitree-go2-basic",
              "examples": [
                "unitree-go2-basic"
              ],
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArgsCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArgsCheck"
                }
              }
            },
            "description": "`{ name, args: [{ tokens, target, error }] }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Check `dimos run <name>` arguments without running anything: what each one sets, or why dimos won't take it",
        "tags": [
          "blueprints"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
    "/dimos/blueprints/{name}/config": {
      "get": {
        "description": "Imports the blueprint in a child process (180 s timeout; cached 10 min) and reads each module's pydantic `config` model: every field a person can set, its type, default, description, whether it's required or inherited from ModuleConfig, its choices (Enum/Literal) and the blueprint's value. A module whose config can't be read has `error` instead of failing the whole answer.",
        "operationId": "get_dimos_blueprints_name_config",
        "parameters": [
          {
            "description": "blueprint name, e.g. unitree-go2-basic",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "description": "blueprint name, e.g. unitree-go2-basic",
              "examples": [
                "unitree-go2-basic"
              ],
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintConfig"
                }
              }
            },
            "description": "`{ name, modules: [{ module, class, args: [{ name, type, default, description, required, base, choices?, value? }], error? }] }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "A blueprint's configurable args per module: name, type, default, description, and the value the blueprint sets (a module that can't be read carries its own error)",
        "tags": [
          "blueprints"
        ],
        "x-agent": true,
        "x-family": "dimos"
      },
      "put": {
        "description": "Replaces config.yaml's `dimos.module_config.<name>` with `overrides` ({module: {field: value}}; null drops a field, a module left empty is dropped, all empty removes the blueprint's entry), after checking each module and field against the blueprint's config (as a launch's `overrides.modules`). A secret sent as \u2022\u2022\u2022 keeps its saved value. 400 for a bad name or a value its field refuses. Answers like GET.",
        "operationId": "put_dimos_blueprints_name_config",
        "parameters": [
          {
            "description": "blueprint name, e.g. unitree-go2-basic",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "description": "blueprint name, e.g. unitree-go2-basic",
              "examples": [
                "unitree-go2-basic"
              ],
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlueprintConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintConfig"
                }
              }
            },
            "description": "`{ name, modules: [...], overrides }`, with the saved module config"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Save Desktop's module config for a blueprint; it becomes `--<module>.<field>=value` on every launch of it",
        "tags": [
          "blueprints"
        ],
        "x-agent": false,
        "x-family": "dimos"
      }
    },
...
    "/dimos/runs": {
      "get": {
        "description": "Every live run on this computer, newest first, whoever started it: dimos's run registry (pid alive), plus runs in other registries (another DIMOS_HOME or XDG_STATE_HOME: a test Desktop, an agent, a terminal) and `dimos run` processes not registered yet; each says where it came from (`registry`, `owner`, `command`, `ours`) and whether POST /dimos/runs/stop can stop it (`stoppable`, `whyNot`). A registry entry whose pid now belongs to a newer process is stale and left out. `seenOnBus`: a dimos run answering on the bus that no listed run accounts for (on another machine through the zenoh connection, or a process this gateway can't see), probed in the background every 10 s; it can't be stopped from here. Also this gateway's last launch with its phase. No side effects.",
        "operationId": "get_dimos_runs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "`{ runs: [{ run_id, pid, blueprint, started_at, log_dir, registry, owner, command, ours, stoppable, whyNot }], launch: Launch | null, seenOnBus: [{ where, peer, note }] }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Running blueprints (run id, blueprint, pid, log_dir) and the launch this gateway started",
        "tags": [
          "runs"
        ],
        "x-agent": true,
        "x-family": "dimos"
      },
      "post": {
        "description": "Starts `dimos [--key value ...] run <blueprint>` in the checkout, in its own session, with Desktop's saved GlobalConfig overrides, then the body's, then `--replay` if asked. Answers at once with phase `starting`; `launch` events (or GET /dimos/runs) follow it to running, stopped or failed. 400 when the checkout's dimos is outside Desktop's range (unless config.yaml `dimos.ignore_version_range`), the name is bad, an override (saved or given) isn't a GlobalConfig flag or valid value, while the last launch is still starting, running or stopping (any process of it left), and while any other dimos run is on this machine (one in dimos's run registry, or a coordinator answering on the bus: two runs share module RPC names, so one's start and stop calls reach the other's modules); 500 when dimos isn't installed or won't start.",
        "operationId": "post_dimos_runs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Launch"
                }
              }
            },
            "description": "`Launch`: `{ blueprint, phase, startedAt, pid, output, runId, logDir, error, overrides, steps: [{ label, state: done|now|todo|failed, detail }], problems: [{ level, text, fix, line }] }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Launch a blueprint (stops nothing; check /dimos/runs first)",
        "tags": [
          "runs"
        ],
        "x-agent": true,
        "x-family": "dimos",
        "x-mcp-tool": "run_blueprint"
      }
    },
...
  },
...
}